+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ New features
 * Add function ``modeling.is_linearized``
 * Add function ``modeling.plot_transformed_eta_distributions``
 * Add function ``modeling.create_config_template``
+* Add function ``modeling.get_dv_symbol``
 
 0.91.0 (2023-03-03)
 -------------------